backend/requirements.txt
Flask==3.0.3
Flask_Bcrypt==1.0.1
Flask_Caching==2.3.0
Flask_RESTful==0.3.10
Flask_Security_Too==5.5.0
flask_sqlalchemy==3.1.1
